emigrate categories context to React Redux
