+++ src/components/sing-up-form/sing-up-form.component.jsx
@@ -5,7 +5,7 @@ import {
 } from "../../utils/firebase/firebase.utils";
 import FormInput from "../form-input/form-input.component";
 import Button from "../button/button.component";
-// import { UserContext } from "../../context/user.context";
+
 import { SingUpContainer } from "./sing-up-form.styles.jsx";
 
 const defaultFormFields = {
@@ -17,9 +17,6 @@ const defaultFormFields = {
 const SingUpForm = () => {
   const [formFields, setFormFields] = useState(defaultFormFields);
   const { displayName, email, password, confirmPassword } = formFields;
-  // const {setCurrentUser}= useContext(UserContext);
-
-  //   console.log(formFields);
 
   const resetFromFields = () => {
     setFormFields(defaultFormFields);
