--- src/context/user.context.jsx
@@ -1,64 +1,0 @@
-import { createContext, useEffect, useReducer } from "react";
-import {
-  onAuthStateChangedListener,
-  createUserDocumentFromAuth,
-} from "../utils/firebase/firebase.utils";
-import { createAction } from "../utils/firebase/reducer/reducer.utils";
-
-// The actual value i want to access
-export const UserContext = createContext({
-  currentUser: null,
-  setCurrentUser: () => null,
-});
-
-// How to use Reducers ********************************************************************
-
-export const USER_ACTION_TYPES = {
-  SET_CURRENT_USER: "SET_CURRENT_USER",
-};
-
-const userReducer = (state, action) => {
-  console.log("Dispatch");
-  console.log(action);
-
-  const { type, payload } = action;
-
-  switch (type) {
-    case USER_ACTION_TYPES.SET_CURRENT_USER:
-      return {
-        ...state,
-        currentUser: payload,
-      };
-
-    default:
-      throw new Error(`Unhandled type ${type} in userReducer`);
-  }
-};
-
-const INITIAL_STATE = {
-  currentUser: null,
-};
-
-// provider
-export const UserProvider = ({ children }) => {
-  const [{ currentUser }, dispatch] = useReducer(userReducer, INITIAL_STATE);
-  console.log(currentUser);
-
-  const setCurrentUser = (user) => {
-    dispatch(createAction(USER_ACTION_TYPES.SET_CURRENT_USER, user));
-  };
-
-  const value = { currentUser, setCurrentUser };
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChangedListener((user) => {
-      // console.log(user);
-      if (user) {
-        createUserDocumentFromAuth(user);
-      }
-      setCurrentUser(user);
-    });
-  }, []);
-
-  return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
-};
