+++ src/routes/categories-preview/categories-preview.component.jsx
@@ -1,33 +1,24 @@
-import { CategoriesContext } from "../../context/categories.context";
+// import { CategoriesContext } from "../../context/categories.context";
 import { Fragment, useContext } from "react";
 import CategoryPreview from "../../components/category-preview/category-preview.component";
+import { useSelector } from "react-redux";
+import { selectCategoriesMap } from "../../store/categories/categories.selector";
 
+const CategoriesPreview = () => {
+  //   const { categoriesMap } = useContext(CategoriesContext);
+  const categoriesMap = useSelector(selectCategoriesMap);
 
+  return (
+    <Fragment>
+      {Object.keys(categoriesMap).map((title) => {
+        const products = categoriesMap[title];
 
-
-const CategoriesPreview = ()=>{
-    const {categoriesMap} = useContext(CategoriesContext);
-
- return (
-   
-<Fragment>
-{
-    Object.keys(categoriesMap).map((title)=> {
-       const products = categoriesMap[title];
-        
-        return <CategoryPreview key={title} title={title} products={products}/>
-    })
-}
-
-</Fragment>
-   
-
- )
-
-
-
-
-}
-
+        return (
+          <CategoryPreview key={title} title={title} products={products} />
+        );
+      })}
+    </Fragment>
+  );
+};
 
 export default CategoriesPreview;
