+++ src/store/categories/categories.action.js
@@ -1,0 +1,5 @@
+import { createAction } from "../../utils/firebase/reducer/reducer.utils";
+import { CATEGORIES_ACTION_TYPES } from "./categories.type";
+
+export const setCategoriesMap = (categoriesMap) =>
+  createAction(CATEGORIES_ACTION_TYPES.SET_CATEGORIES_MAP, categoriesMap);
