+++ src/store/categories/categories.reducer.js
@@ -1,0 +1,19 @@
+import { CATEGORIES_ACTION_TYPES } from "./categories.type";
+
+export const CATEGORIES_INITIAL_STATE = {
+  categoriesMap: {},
+};
+
+export const categoriesReducer = (
+  state = CATEGORIES_INITIAL_STATE,
+  action = {}
+) => {
+  const { type, payload } = action;
+
+  switch (type) {
+    case CATEGORIES_ACTION_TYPES.SET_CATEGORIES_MAP:
+      return { ...state, categoriesMap: payload };
+    default:
+      return state;
+  }
+};
