+++ src/store/categories/categories.selector.js
@@ -1,0 +1,1 @@
+export const selectCategoriesMap = (state) => state.categories.categoriesMap;
