+++ src/store/categories/categories.type.js
@@ -1,0 +1,3 @@
+export const CATEGORIES_ACTION_TYPES = {
+  SET_CATEGORIES_MAP: "category/SET_CATEGORIES_MAP",
+};
